fix test break upon changing int to double

+++ src/main/java/models/Event.java
@@ -8,7 +8,7 @@ public class Event {
     private int foodCost;
     private int drinkCost;
     private int entertainmentCost;
-    private double eventCost;
+    private int eventCost;
 
     public void setGuestCount(int count) {
         this.guestCount = count;
@@ -88,7 +88,7 @@ public class Event {
         this.eventCost = this.foodCost + this.drinkCost + this.entertainmentCost;
     }
 
-    public double getEventCost() {
+    public int getEventCost() {
         return eventCost;
     }
